organize filequeue, add consumer test
